Move filter utils to metabase-lib
